Move factory method to a dedicated files as they are used by tests also.
