+++ libraries/mediaviewer/impl/src/main/kotlin/io/element/android/libraries/mediaviewer/impl/gallery/ui/MediaItemAudioProvider.kt
@@ -9,11 +9,8 @@ package io.element.android.libraries.mediaviewer.impl.gallery.ui
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import io.element.android.libraries.core.preview.loremIpsum
-import io.element.android.libraries.matrix.api.core.EventId
-import io.element.android.libraries.matrix.api.core.UniqueId
-import io.element.android.libraries.matrix.api.media.MediaSource
-import io.element.android.libraries.mediaviewer.api.anAudioMediaInfo
 import io.element.android.libraries.mediaviewer.impl.model.MediaItem
+import io.element.android.libraries.mediaviewer.impl.model.aMediaItemAudio
 
 class MediaItemAudioProvider : PreviewParameterProvider<MediaItem.Audio> {
     override val values: Sequence<MediaItem.Audio>
@@ -27,21 +24,4 @@ class MediaItemAudioProvider : PreviewParameterProvider<MediaItem.Audio> {
                 caption = loremIpsum,
             ),
         )
-}
-
-fun aMediaItemAudio(
-    id: UniqueId = UniqueId("fileId"),
-    eventId: EventId? = null,
-    filename: String = "filename",
-    caption: String? = null,
-): MediaItem.Audio {
-    return MediaItem.Audio(
-        id = id,
-        eventId = eventId,
-        mediaInfo = anAudioMediaInfo(
-            filename = filename,
-            caption = caption,
-        ),
-        mediaSource = MediaSource(""),
-    )
 }
