+++ libraries/mediaviewer/impl/src/main/kotlin/io/element/android/libraries/mediaviewer/impl/gallery/ui/MediaItemFileProvider.kt
@@ -9,11 +9,8 @@ package io.element.android.libraries.mediaviewer.impl.gallery.ui
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import io.element.android.libraries.core.preview.loremIpsum
-import io.element.android.libraries.matrix.api.core.EventId
-import io.element.android.libraries.matrix.api.core.UniqueId
-import io.element.android.libraries.matrix.api.media.MediaSource
-import io.element.android.libraries.mediaviewer.api.aPdfMediaInfo
 import io.element.android.libraries.mediaviewer.impl.model.MediaItem
+import io.element.android.libraries.mediaviewer.impl.model.aMediaItemFile
 
 class MediaItemFileProvider : PreviewParameterProvider<MediaItem.File> {
     override val values: Sequence<MediaItem.File>
@@ -27,21 +24,4 @@ class MediaItemFileProvider : PreviewParameterProvider<MediaItem.File> {
                 caption = loremIpsum,
             ),
         )
-}
-
-fun aMediaItemFile(
-    id: UniqueId = UniqueId("fileId"),
-    eventId: EventId? = null,
-    filename: String = "filename",
-    caption: String? = null,
-): MediaItem.File {
-    return MediaItem.File(
-        id = id,
-        eventId = eventId,
-        mediaInfo = aPdfMediaInfo(
-            filename = filename,
-            caption = caption,
-        ),
-        mediaSource = MediaSource(""),
-    )
 }
