+++ libraries/mediaviewer/impl/src/main/kotlin/io/element/android/libraries/mediaviewer/impl/gallery/ui/MediaItemVideoProvider.kt
@@ -8,10 +8,8 @@
 package io.element.android.libraries.mediaviewer.impl.gallery.ui
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import io.element.android.libraries.matrix.api.core.UniqueId
-import io.element.android.libraries.matrix.api.media.MediaSource
-import io.element.android.libraries.mediaviewer.api.aVideoMediaInfo
 import io.element.android.libraries.mediaviewer.impl.model.MediaItem
+import io.element.android.libraries.mediaviewer.impl.model.aMediaItemVideo
 
 class MediaItemVideoProvider : PreviewParameterProvider<MediaItem.Video> {
     override val values: Sequence<MediaItem.Video>
@@ -21,20 +19,4 @@ class MediaItemVideoProvider : PreviewParameterProvider<MediaItem.Video> {
                 duration = null,
             ),
         )
-}
-
-fun aMediaItemVideo(
-    id: UniqueId = UniqueId("videoId"),
-    mediaSource: MediaSource = MediaSource(""),
-    duration: String? = "1:23",
-): MediaItem.Video {
-    return MediaItem.Video(
-        id = id,
-        eventId = null,
-        mediaInfo = aVideoMediaInfo(
-            duration = duration
-        ),
-        mediaSource = mediaSource,
-        thumbnailSource = null,
-    )
 }
